Upload of 24 Hour Build
This was the final build made by us during the game jam.

The controls are:
W + S/Up and Down for moving the spaceship.
Spacebar to fire a projectile.

The goal is to destroy the enemy ships. The game ends when you collide three times with an enemy ship. Try to get a high score!